requirements.txt
numpy>=1.21.0
pandas>=1.3.0
geopandas>=0.9.0
shapely>=1.7.1
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
imbalanced-learn>=0.8.0
xgboost>=1.5.0
folium>=0.12.0
joblib>=1.0.0
pillow>=8.0.0
opencv-python>=4.5.0